+++ packages/components/src/components/tag/tag-utils.spec.ts
@@ -22,7 +22,7 @@ const colorsWithThemeCombinations: [TagColor, Theme, boolean][] = [
   ['notification-success', 'light', false],
   ['notification-warning', 'light', false],
   ['notification-error', 'light', false],
-  ['notification-neutral', 'light', false], // 'notification-neutral' is deprecated (replaced with 'notification-information')
+  ['notification-neutral', 'light', false], // 'notification-neutral' is deprecated (replaced with 'notification-info')
   ['notification-info', 'light', false],
   ['background-surface', 'dark', false],
   ['background-default', 'dark', false],
@@ -30,7 +30,7 @@ const colorsWithThemeCombinations: [TagColor, Theme, boolean][] = [
   ['notification-success', 'dark', false],
   ['notification-warning', 'dark', false],
   ['notification-error', 'dark', false],
-  ['notification-neutral', 'dark', false], // 'notification-neutral' is deprecated (replaced with 'notification-information')
+  ['notification-neutral', 'dark', false], // 'notification-neutral' is deprecated (replaced with 'notification-info')
   ['notification-info', 'dark', false],
 ];