+++ packages/components-js/tests/vrt/puppeteer/specs/component-scaling.vrt.ts
@@ -5,15 +5,17 @@ import {
 } from '@porsche-design-system/shared/testing';
 
 const components = [
-  'checkbox-wrapper',
-  'radio-button-wrapper',
-  'link',
   'button',
-  'tag-dismissible',
-  'spinner',
+  'checkbox-wrapper',
+  'link',
   'popover',
-  'text-field-wrapper',
+  'radio-button-wrapper',
   'scroller',
+  'spinner',
+  'tag',
+  'tag-dismissible',
+  'text-field-wrapper',
+  'textarea-wrapper',
 ];
 
 it.each(components)('should have no visual regression for scaled component %s', async (component) => {