+++ packages/components/CHANGELOG.md
@@ -41,6 +41,34 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 + require('@porsche-design-system/components-js/partials').getIconLinks({ icons: ['arrow-right'] })
 ```
 
+- Color `notification-neutral` for `Tag` is deprecated now, replaced by `notification-info`
+
+```diff
+- <p-tag color="notification-neutral">Color label</p-tag>
++ <p-tag color="notification-info">Color label</p-tag>
+```
+
+- Color `neutral-contrast-high` for `Tag` is deprecated now, replaced by `primary`
+
+```diff
+- <p-tag color="neutral-contrast-high">Color label</p-tag>
++ <p-tag color="primary">Color label</p-tag>
+```
+
+- Color `background-default` for `Tag` is deprecated now, replaced by `background-base`
+
+```diff
+- <p-tag color="background-default">Color label</p-tag>
++ <p-tag color="background-base">Color label</p-tag>
+```
+
+- Color `background-default` for `Tag Dismissible` is deprecated now, replaced by `background-base`
+
+```diff
+- <p-tag-dismissible color="background-default">Color label</p-tag>
++ <p-tag-dismissible color="background-base">Color label</p-tag>
+```
+
 - Default value of prop `name` of `Icon` has changed from **"arrow-head-right"** to **"arrow-right"**
 - Default value of prop `variant` of `Link` and `Button` has changed from **"secondary"** to **"primary"**
 - Default value of prop `icon` of `Link` and `Button` has changed from **"arrow-head-right"** to **"none"**. Therefore,
@@ -61,7 +89,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ```
 
 - `Spinner`, `Icon`, `Link Pure`, `Button Pure`, `Link`, `Button`, `Checkbox Wrapper`, `Radio Button Wrapper`,
-  `Popover`, `Tag Dismissible`, `Textarea Wrapper` and `Text Field Wrapper` match new design language
+  `Popover`, `Tag`, `Tag Dismissible`, `Textarea Wrapper` and `Text Field Wrapper` matches new design language
 - CSS global variables names
 
 ```diff